+++ trl/trainer/__init__.py
@@ -38,9 +38,10 @@ if is_diffusers_available():
 
 from .dpo_trainer import DPOTrainer
 from .iterative_sft_trainer import IterativeSFTTrainer
+from .kto_config import KTOConfig
 from .kto_trainer import KTOTrainer
 from .ppo_config import PPOConfig
 from .ppo_trainer import PPOTrainer
 from .reward_trainer import RewardTrainer, compute_accuracy
 from .sft_trainer import SFTTrainer
-from .training_configs import KTOConfig, RewardConfig
+from .training_configs import RewardConfig
